% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_S3.R
\name{coef.ICS-S3}
\alias{coef.ICS-S3}
\alias{coef.ICS}
\title{Extract the Coefficient Matrix of the ICS Transformation}
\usage{
\method{coef}{ICS}(object, select = NULL, drop = FALSE, index = NULL, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"ICS"} containing
results from an ICS transformation.}

\item{select}{an integer, character, or logical vector specifying for which
components to extract the coefficients, or \code{NULL} to extract the
coefficients for all components.}

\item{drop}{a logical indicating whether to return a vector rather than a
matrix in case coefficients are extracted for a single component (defaults
to \code{FALSE}).}

\item{index}{an integer vector specifying for which components to extract
the coefficients, or \code{NULL} to extract coefficients for all components.
Note that \code{index} is deprecated and may be removed in the future, use
\code{select} instead.}
}
\value{
A numeric matrix or vector containing the coefficients for the
requested components.
}
\description{
Extract the coefficient matrix of a linear transformation to an invariant
coordinate system. Each row of the matrix contains the coefficients of the
transformation to the corresponding component.
}
\seealso{
\code{\link{ICS}()}

\code{\link{gen_kurtosis}()}, \code{\link{components}()},
\code{\link[=fitted.ICS]{fitted}()}, and \code{\link[=plot.ICS]{plot}()}
methods
}
\author{
Andreas Alfons and Aurore Archimbaud
}
