% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covW.R
\name{covW}
\alias{covW}
\title{One-step M-estimator covW}
\usage{
covW(X, na.action = na.fail, alpha = 1, cf = 1)
}
\arguments{
\item{X}{numeric nxp data matrix or dataframe.}

\item{na.action}{a function which indicates what should happen when the data contain 'NA's. Default is to fail.}

\item{alpha}{parameter of the one-step M-estimator. By default equals to 1.}

\item{cf}{consistency factor of the one-step M-estimator. By default equals to 1}
}
\description{
One-step M-estimator covW
}
