% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_S3.R
\name{gen_kurtosis}
\alias{gen_kurtosis}
\alias{gen_kurtosis.ICS}
\title{Extract the Generalized Kurtosis Values of the ICS Transformation}
\usage{
gen_kurtosis(object, ...)

\method{gen_kurtosis}{ICS}(object, select = NULL, scale = FALSE, index = NULL, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"ICS"} containing
results from an ICS transformation.}

\item{\dots}{additional arguments to be passed down.}

\item{select}{an integer, character, or logical vector specifying for which
components to extract the generalized kurtosis values, or \code{NULL} to
extract the generalized kurtosis values of all components.}

\item{scale}{a logical indicating whether to scale the generalized kurtosis
values to have product 1 (defaults to \code{FALSE}).}

\item{index}{an integer vector specifying for which components to extract
the generalized kurtosis values, or \code{NULL} to extract the generalized
kurtosis values of all components.  Note that \code{index} is deprecated
and may be removed in the future, use \code{select} instead.}
}
\value{
A numeric vector containing the generalized kurtosis values of the
requested components.
}
\description{
Extract the generalized kurtosis values of the components obtained via an
ICS transformation.
}
\seealso{
\code{\link{ICS}()}

\code{\link[=coef.ICS]{coef}()}, \code{\link{components}()},
\code{\link[=fitted.ICS]{fitted}()}, and \code{\link[=plot.ICS]{plot}()}
methods
}
\author{
Andreas Alfons and Aurore Archimbaud
}
