\name{ics.components}
\alias{ics.components}

\title{ Extracting ICS Components}
\description{
  Function to extract the ICS components of a \code{ics} object.
}
\usage{
ics.components(object)
}

\arguments{
  \item{object}{object of class \code{ics}.}
}


\value{
 Dataframe that contains the components.
}
\author{Klaus Nordhausen, \email{klaus.nordhausen@uta.fi}}

\seealso{\code{\link{ics-class}} and \code{\link{ics}}}


\keyword{models}
\keyword{multivariate}

