% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_S3.R
\name{ICS_scatter}
\alias{ICS_scatter}
\alias{ICS_cov}
\alias{ICS_cov4}
\alias{ICS_covW}
\alias{ICS_covAxis}
\title{Location and Scatter Estimates for ICS}
\usage{
ICS_cov(x, location = TRUE)

ICS_cov4(x, location = c("mean3", "mean", "none"))

ICS_covW(x, location = TRUE, alpha = 1, cf = 1)

ICS_covAxis(x, location = TRUE)
}
\arguments{
\item{x}{a numeric matrix or data frame.}

\item{location}{for \code{ICS_cov()}, \code{ICS_covW()}, and
\code{ICS_covAxis()}, a logical indicating whether to include the sample
mean as location estimate (defaults to \code{TRUE}).  For \code{ICS_cov4()},
alternatively a character string specifying the location estimate can be
supplied.  Possible values are \code{"mean3"} for a location estimate based
on third moments (the default), \code{"mean"} for the sample mean, or
\code{"none"} to not include a location estimate.}

\item{alpha}{parameter of the one-step M-estimator (defaults to 1).}

\item{cf}{consistency factor of the one-step M-estimator (defaults to 1).}
}
\value{
An object of class \code{"ICS_scatter"} with the following
components:
\item{location}{if requested, a numeric vector giving the location
estimates.}
\item{scatter}{a numeric matrix giving the scatter matrix.}
\item{label}{a character string providing a label for the scatter matrix.}
}
\description{
Compute a scatter matrix and an optional location vector to be used in
transforming the data to an invariant coordinate system or independent
components.
}
\details{
\code{ICS_cov()} is a wrapper for the sample covariance matrix as computed
by \code{\link[stats]{cov}()}.

\code{ICS_cov4()} is a wrapper for the scatter matrix based on fourth
moments as computed by \code{\link{cov4}()}. Although the returned location
estimate can be specified to be based on third moments as computed by
\code{\link{mean3}()}, the scatter matrix of fourth moments is always
computed with respect to the sample mean.

\code{ICS_covW()} is a wrapper for the one-step M-estimator of scatter as
computed by \code{\link{covW}()}.

\code{ICS_covAxis()} is a wrapper for the one-step Tyler shape matrix as
computed by \code{\link{covAxis}()}, which is can be used to perform
Principal Axis Analysis.
}
\references{
Critchley, F., Pires, A. and Amado, C. (2006) Principal Axis Analysis.
Technical Report, \bold{06/14}. The Open University, Milton Keynes.

Oja, H., Sirkia, S. and Eriksson, J. (2006) Scatter Matrices and Independent
Component Analysis. \emph{Austrian Journal of Statistics}, \bold{35}(2&3),
175-189. \doi{10.17713/ajs.v35i2&3.364}.

Tyler, D.E., Critchley, F., Duembgen, L. and Oja, H. (2009) Invariant
Co-ordinate Selection. \emph{Journal of the Royal Statistical Society,
Series B}, \bold{71}(3), 549--592. \doi{10.1111/j.1467-9868.2009.00706.x}.
}
\seealso{
\code{\link{ICS}()}

\code{\link[base:colSums]{colMeans}()}, \code{\link{mean3}()}

\code{\link[stats]{cov}()}, \code{\link{cov4}()}, \code{\link{covW}()},
\code{\link{covAxis}()}
}
\author{
Andreas Alfons and Aurore Archimbaud
}
