\name{print.ics}
\docType{methods}
\alias{print.ics}
%\alias{show-method}
\alias{show,ics-method}
\title{Basic information of ICS Object}
\description{
  Prints the minimal information of an \code{ics} object.
}
\usage{
\S4method{show}{ics}(object)
}

\arguments{
  \item{object}{object of class \code{ics}. }
}

\author{Klaus Nordhausen}

\seealso{\code{\link{ics-class}} and \code{\link{ics}}}

\keyword{methods}
\keyword{print}

