\name{print.ics2}
\docType{methods}
\alias{print.ics2}
%\alias{show-method}
\alias{show,ics2-method}
\title{Basic information of ICS2 Object}
\description{
  Prints the minimal information of an \code{ics2} object.
}
\usage{
\S4method{show}{ics2}(object)
}

\arguments{
  \item{object}{object of class \code{ics2}. }
}

\author{Klaus Nordhausen}

\seealso{\code{\link{ics2-class}} and \code{\link{ics2}}}

\keyword{methods}
\keyword{print}
