% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICS_S3.R
\name{summary.ICS-S3}
\alias{summary.ICS-S3}
\alias{summary.ICS}
\title{To summarize an \code{ICS} object}
\usage{
\method{summary}{ICS}(object, ...)
}
\arguments{
\item{object}{object of class \code{ICS}.}

\item{...}{additional arguments passed to \code{\link[=print.ICS]{print.ICS()}}.}
}
\description{
Summarizes and prints an \code{ICS} object in an informative way.
}
\examples{
data("iris")
X <- iris[,-5]
out <- ICS(X)
summary(out)
}
\seealso{
\code{\link{ICS}()}

\code{\link[=print.ICS]{print.ICS()}}
}
\author{
Andreas Alfons and Aurore Archimbaud
}
