\name{summary.ics}
\docType{methods}
\alias{summary.ics}
%\alias{summary-method}
\alias{summary,ics-method}

\title{To summarize an ICS object}
\description{
  Summarizes and prints a \code{ics} object in an informative way.
}
\usage{
\S4method{summary}{ics}(object, digits = 4)
}

\arguments{
  \item{object}{object of class \code{ics}.}
  \item{digits}{ number of digits for the numeric output.}
}


\author{Klaus Nordhausen}

\seealso{\code{\link{ics-class}} and \code{\link{ics}}}

\keyword{methods}
\keyword{print}
