\name{summary.ics2}
\docType{methods}
\alias{summary.ics2}

\alias{summary,ics2-method}

\title{To summarize an ICS2 object}
\description{
  Summarizes and prints a \code{ics2} object in an informative way.
}
\usage{
\S4method{summary}{ics2}(object, digits = 4)
}

\arguments{
  \item{object}{object of class \code{ics2}.}
  \item{digits}{ number of digits for the numeric output.}
}


\author{Klaus Nordhausen}

\seealso{\code{\link{ics2-class}} and \code{\link{ics2}}}

\keyword{methods}
\keyword{print}
