\name{coef.ics}
\docType{methods}
\alias{coef.ics}
\alias{coef-method}
\alias{coef,ics-method}

\title{To extract the Unmixing Matrix}
\description{
  Extracts the unmixing matrix of a class \code{ics} object.
}
\usage{
\S4method{coef}{ics}(object)
}
\arguments{
  \item{object}{object of class \code{ics}.}
}
\value{
  The unmixing matrix of a class \code{ics} object.
}

\author{Klaus Nordhausen}

\seealso{\code{\link{ics-class}} and \code{\link{ics}}}

\keyword{models}
\keyword{multivariate}
\keyword{methods}
